Migrate App to typescript
